Don’t leave behind pending commands on command failure
When the user runs a set of commands (not in the background), we grab up
to fifty planned commands and mark them as “pending” so that nobody
running in parallel will try to run the same commands as us. As each
command runs, we replace that pending command with a finish, usually a
success. However, if there is a failure that aborts the whole batch run
(e. g. “maxlag exceeded” or “user blocked”), we failed to reset the
pending commands to the “planned” state, leaving behind up to 50
commands that would now never be run.

This is fixed by introducing a method to BatchCommandRecords that’s the
reverse of make_plans_pending (uninventively named make_pendings_planned
– perhaps I’ll rename it to something better in the future), turning all
the commands that are still pending back into planned ones, and call
that in a “finally” clause. (We also do it in the background runner,
though there a problem would only materialize if some exception occurred
– a “maxlag exceeded” failure would be stored, overriding the one
pending command that the background runner grabs at a time.)