Catch error if user account does not exist
If a user tries to view a batch for a wiki where they do not have a
local account yet, the MediaWiki API will respond with an error
mwoauth-invalid-authorization-invalid-user when we try to find out the
user’s capabilities. Catch this error and treat the user as anonymous on
the local wiki in that case, but ask Meta for their global user ID so we
can still allow stewards to stop such background runs for such batches.