Let can_continue_batch() return a datetime too
Some command failures (chiefly “maxlag exceeded”) should suspend a
background while for a while, but don’t need to stop it completely. Let
the can_continue_batch() function return a datetime in addition to a
bool to indicate this. Callers don’t make use of this information yet
(though they still need to be updated because otherwise a datetime is
treated as True whereas we want it to act like False).