Test make_plan_pending_background for all stores
When I rearranged the store tests a while back, this test didn’t work on
the in-memory store yet, so even though it doesn’t use any database
specifics, it stayed in the database tests for that time. This has been
fixed in the previous three commits, so now we can run the test on the
in-memory store as well and ensure it stays functional.