Add first/last links to batch page
The first page of commands of a batch is easy: the one with offset 0.
The last page of commands is a bit trickier, because we have two
options: if a batch has, say, 173 commands, then (at a limit of 50) does
the last page start at offset 150 or 123? And what if the user is
currently on offset 111, so that clicking “next” will take them to yet
another last page?

A last page link is too useful to be blocked by such questions, though,
so I picked one: the last page starts at offset 150, so that the offsets
stay nice and round (though users can specify others by hand). Seeing a
list with less than 50 (limit) commands on it also emphasizes that this
is the last page of commands of a batch.

The length of the commands is extracted into a variable so we don’t run
the same SQL query twice (it’s not cached elsewhere).