Add optional title to batches
Harmonia Amanda suggested that users should be able to add a title to
batches, which would be used in edit summaries to explain what this
batch does beyond the raw category changes. This adds the field to the
batch classes and teaches the stores to store it; support for users to
define a title, and using the title for display and summaries, will be
implemented next.

In the database, the title is normalized because it’s likely that many
batches will use the same title, according to Harmonia – generic titles
like “deleting 'People of' since already in 'Births in'” (which wouldn’t
mention the specific categories in the batch title because many
categories would be collected in the same batch).