Cache parsed summaries
On the same domain, the same edit summary should always parse to the
same HTML, so we can cache it for a long time. (It would become invalid
if a link turns from a bluelink to a redlink or vice versa, for example,
but meh, I don’t think we need to worry about that very much. Worst
case, just restart the webserver to flush the cache.)