Extract internet_test decorator
I originally intended to use this for parse_wikitext tests before
realizing that I can just inject FakeSession objects there, as when
testing siteinfo. It’s still a useful decorator, though.

The stub for the decorator module is a nasty piece of work. I *think*
it’s correct, but even if that’s true it’s not even complete, due to
type system limitations (the various ... should relate to each other and
ideally also to the args and kwargs types). But at least this ensemble
currently passes mypy, that’s good enough for me right now.