Parse titles as summaries using an anonymous session
I *think* the parse result of a summary is always independent of the
user requesting it, but I’m not quite confident enough. If it isn’t,
user-specific information could leak through the summary cache in
parse_wikitext, and since one of the frequent users of the tool
(Strakhov) is an admin on Commons, I’m not willing to risk this leak.