Add BatchCommandRecords.get_summary()
This returns the number of commands for each different status, or type,
and will be displayed on top of a batch page for a quick overview.

On the database backend, the required query is nicely efficient because
it can make optimal use of the command_batch_status index, which for
this query is a covering one.

Perhaps I’ll later update this to also include information on how many
commands were retried (à la “5 maxlag exceeded, all retried”), but
that’ll make the signature and query more complicated, so for now this
should do just fine.