Add consistent titles to all command status badges
In the summary, it’s weird to have two kinds of “done” next to each
other when they’re not hyperlinks to Special:PermanentLink /
Special:Diff pages to show which is which. (In fact, I’m not sure how
clear that was even with the links, so this might help in the regular
command list as well.)

Repeatable failures say “retrying later”, even though strictly speaking
that’s not true for some early batches (before retry was implemented); I
think it’s more important to include this information in recent batches,
though, so for now I can live with this slight inaccuracy.