Add and link labels for page parts
On the index page, link the existing “latest batches” heading to the
table below it via aria-labelledby, and add a new “new batch” heading
likewise linked to the form. This hopefully mitigates the “user is
dropped into the form with no context” downside of the autofocus
attribute added in the previous commit, though the information I can
find on this is apparently unclear on whether this actually helps.