Fix relative links in edit summaries
The HTML returned by the MediaWiki API includes relative links, so we
have to fix them up by prepending the wiki’s host to the href, using the
Beautiful Soup library to parse the HTML first. The fix_markup()
function is adapted from SpeedPatrolling, with the most significant
difference being that the host is no longer hard-coded.