Extract timestamp module
This holds the _now() function (now without underscore), which
previously sat awkwardly in store, and the conversion to and from
integer UTC timestamps, which previously was in DatabaseStore for no
good reason (it could’ve been a module global all along).

This, like the previous commit, is motivated by the intention to track
query runtimes in a separate module, which will also require the
timestamp-related functions but shouldn’t depend on database.