Add querytime module to track query execution times
This module introduces a new cursor class that can be used to track
query execution times in a separate querytime table (with the query text
normalized via another table, querytext), to figure out which queries
might need to be optimized. Data is flushed to the database periodically
(but this shouldn’t happen too often, because the flushing also involves
queries that are timed).

This is not hooked up to anything in the main app yet.