Rearrange foreground/background run buttons
There have been several batches where only the first 50 commands were
run – users clicked the first button (run in foreground), ran the first
page of commands, and never noticed that the rest wasn’t done. To avoid
this, emphasize the background run button a lot more on the first page,
and add a warning to the foreground run button when not all commands are
shown.