Add actions for manipulating sort keys
This adds three kinds of actions with sort key when adding a category,
and one kind of action with sort key when removing a category, in
addition to the existing actions (which remain unchanged). This doesn’t
do anything yet – parse_tpsv support is missing (though you can see the
planned syntax in the str() of the actions) – but that will be resolved
by the next commit.