Add sitematrix module to turn dbnames into domains
I’m planning to add integration with PagePile [1], but that doesn’t
directly return wiki domains, so we’ll need this module to turn the
dbnames it returns into domains we can use.

In theory, the sitematrix API supports continuation, but when used with
mwapi’s continuation=True it also raises a warning (unrecognized
parameter “continue” because it’s called “smcontinue”), and the default
and max limit of 5000 is more than enough for Wikimedia’s current 934
sites, so for now we don’t use continuation at all and assume that we
get all the data in a single request. However, we can detect when
continuation would become necessary, and warn in that case.

[1]: https://tools.wmflabs.org/pagepile/