Reduce height of commands input on index page
People don’t usually write these completely by hand, so I think it’s
fine to give it less space on the page (most browsers let you resize it
anyways). With this change, the whole form including the submit button
fits neatly on a 720p screen, and on 1080p you can see several of the
latest batches as well.