Allow creating batches from PagePiles
In this mode, the list of pages (and target wiki) comes from a PagePile
ID, and a single list of actions applies to all wikis equally.

This isn’t linked from the index page yet, and I think I might actually
keep it this way, because it’s a somewhat niche feature and also much
more useful as a link target from the PagePile tool itself (I’ll send a
pull request to that effect).