Add create_pagepile() function to pagepile module
This will be used to export batches to PagePiles. The function accepts
streams of titles, and in theory we could stream it to PagePile as well
using requests_toolbelt [1], but it doesn’t look like the current
version of PagePile would handle that correctly (though I haven’t tried
it out), so for now we materialize the whole stream and accept that this
may not work so well for very large piles.

A test is included, but since there’s no test instance of PagePile we
could run it against, it’s disabled by default so that we don’t spam
PagePile with new piles every time the test suite is run.

[1]: https://toolbelt.readthedocs.io/en/latest/uploading-data.html