Show number of notifications in authentication area
If a user gets pinged about the edits they make via this tool, it’s good
if they can see that notification in this tool as well.

On small and medium viewports, we hid the “notifications” text to save
some space – after you follow the link for the first time, it should be
fairly clear what it means.