Copy first/previous/next/last nav below lists too
When you’ve reached the bottom of a list (of batches or command
records), it’s inconvenient to have to scroll all the way back to the
top to go to the next page. Luckily, Jinja2 has block assignments since
version 2.8, so it’s fairly simple to copy the whole <nav> without
duplicating the code. The only slight complication is that we need to
wrap the list of command records in a div with some margin at the bottom
so that the lower navbar isn’t squashed right next to it.