Add preference for default domain
Many users only edit one project with this tool most of the time; with
this preference, that project is pre-filled on the index page as the
target wiki of a new batch. In that case, we can also move the autofocus
to the commands for the new batch instead. (It doesn’t seem very useful
to autofocus the optional title/summary.)

This preference was also suggested by Harmonia Amanda.