Add preference for suggested domains
This lets users customize the datalist of domains suggested on the index
page; we fall back to the old default if the preference is not
specified. (I’m not sure how useful this is, since at least my browser
seems to prioritize my own inputs as suggestions over the datalist
specified in the HTML, but it’s not too hard to implement anyways.)

This preference was also suggested by Harmonia Amanda.