Add proper CSRF error page
The previous behavior wasn’t really acceptable – this new page includes
some advice to users on how to deal with the error. (The soft reload
is necessary to refresh the csrf_token hidden input.) The text is
general enough that it should also be suitable for the cookiecutter
template, but let’s see how it fares with QuickCategories users first.