Filter Beautiful Soup warnings about parsing URLs
Some users use URLs for batch titles. Those stay unchanged when parsing
them for summaries, and when we then pass them into Beautiful Soup, it
warns us that it’s not an HTTP client, and we shouldn’t expect it to
fetch the contents of the URL. Since that’s not our intention, silence
the warning.