Revert "Temporarily disable mypy in Makefile"
The typeshed update was shipped in mypy v0.750.

Unfortunately, in the time since we disabled mypy, a newer version of it
seems to have become more strict about union types. Typeshed marks most
pymysql methods as returning results using the “Gen” type, an alias for
the union of a tuple of Any and a dict of Any, since pymysql can return
either tuples or dicts depending on the cursorclass. Previously, our
code which used only the “tuple” side was accepted by mypy, but now it
complains – my best guess is that it was previously happy if any branch
of a union type matched, but now requires all of them to match. As far
as I can tell, this makes the typeshed annotations for pymysql
completely unusable, and I went for the sledgehammer solution of
replacing them with stubs of my own which only return tuples.

A better solution would probably be to make the connection type generic,
depending on the cursorclass. This might also solve the cast of the
cursor that’s currently necessary in one place, to make the
fetchall_unbuffered() method available. But I can’t be bothered to put
much time into this at the moment.

This also switches the Python version in Travis from 3.7.0a2 to 3.7-dev.
With 3.7.0a2, mypy fails to launch:

ImportError: /home/travis/virtualenv/python3.7.0a2/lib/python3.7/site-packages/97d839f88ac1187b7870__mypyc.cpython-37m-x86_64-linux-gnu.so: undefined symbol: _PyObject_LookupAttrId

3.7-dev works, though. 🤷

This reverts commit a0af4a74d96c4a3e0642cd1dbf69adbab4adfbbc.