Extract Page object from Command
The page of a Command is no longer directly the title string, but rather
an instance of the new Page class. Right now, a Page is identified only
by its title string, but the plan is to later add information there
whether this page should be resolved if it’s a redirect (i. e., edit the
redirect target’s categories) or not (categorize the redirect).

However, the Page also provides a convenient place for the Runner to
store the results of “resolving” a Page (title) via the MediaWiki API,
to a page ID, revision ID, wikitext, and so on. So that information is
now attached to the Page object, and the Runner’s prepared_pages are no
longer necessary.