Add flag whether to follow redirects when resolving pages
Previously, the tool did not resolve redirects when editing pages. This
was most likely not what the user wanted, so introduce a flag on the
Page object: True to resolve redirects, False to not resolve them, and
None to indicate that it was not specified at batch creation time (old
batches). When parsing pages, a “!” prefix for the page title means to
*not* resolve redirects (i. e. the default is to resolve them); however,
for backwards compatibility, when processing old batches, a None flag is
interpreted to *not* resolve redirects (i. e. the batches were created
using syntax that *now* would mean resolving redirects, but preserve
their behavior prior to this change, which would have been *not* to
resolve them).

The database migration maintenance script first declares the new column
with an explicit default of NULL, then sets it to NULL in all existing
rows explicitly, and then redeclares the column without an explicit
default. I don’t think this actually makes a difference in SQL, but I
would like for the database to scream at me if I don’t specify a value,
even if that’s not actual SQL semantics (the DEFAULT NULL is actually
implicit, I believe, unless the column is declared NOT NULL).

The runner tests are significantly expanded to test this, with some
corresponding new pages on Test Wikipedia to test redirect behavior.