Add None return type to all constructors
Mypy, even in --disallow-untyped-defs mode (which I want to make the
default soon), allows constructors to omit the return type as long as
they have at least one (non-self) parameter that is type annotated (to
opt into type checking, which is not done on untyped declarations). We
have one constructor without other parameters, though (InMemoryStore),
and it seems odd to add a None return type only there, so let’s just
bite the bullet and add it on all constructors.

Done with the following command:

    sed -i '/def __init__/ s/):$/) -> None:/' *.py