Cast flask.request.endpoint to str
In typeshed, Request’s endpoint attribute is declared to return
Optional[str] (which is technically correct, since it’s None if matching
failed), but since python/typeshed#3888 [1], the url_for argument must
be a str, non-Optional. In this case, the endpoint will never be None,
so cast it to str.

[1]: https://github.com/python/typeshed/pull/3888