Add read_only_mode feature
The configuration file can now specify a read_only_mode feature, and the
tool will no longer try to make any writes to its database (nor edits on
any wiki, because that would require updating the command status in the
database). Note that a handful of POSTs are still permitted – there’s no
need to disallow creating PagePiles or editing your settings.

This is needed for some upcoming maintenance to ToolsDB [1] – without
read-only mode in the tool, it would just crash trying to write to the
database, I assume. This didn’t take much effort and is much nicer.

[1]: https://lists.wikimedia.org/pipermail/cloud-announce/2020-October/000330.html