Add error handler for DB errors
Add a generic error handler in case of a pymysql OperationalError (could
not connect, connection reset, etc.), and an optional config setting for
an expected_database_error (such as [1]). Unlike the read_only_reason,
this setting is only used if we actually can’t access the database – I
don’t want to accidentally leave the tool down for hours because I
forget to remove the setting even after ToolsDB is back up.

[1]: https://lists.wikimedia.org/pipermail/cloud-announce/2020-December/000347.html