Better workaround for parse API error
Apparently the API error we got was because MediaWiki now tried to parse
the “text” parameter using the default title, and on Wikidata, the
default title, “API”, is in the item namespace, so the default text, the
empty string, is not valid content. Avoid this by using an arbitrary
title in the Special namespace instead, where empty text is always
valid content.

The old fallback for API errors remains in place – after all, it still
makes sense in case of other API errors – but we no longer expect it to
happen.

Bug: T279585