Add margin-bottom to <main>
Otherwise, it’s possible for the end of the page to be flush with the
bottom of the screen, which isn’t very nice.

Most of the time, this didn’t actually happen, but only accidentally:
the pagination buttons are a <ul>, which Bootstrap adds a margin-bottom
to (which isn’t reset in ul.pagination even though it’s not visually
presented as a list). So the problem being fixed here mainly happened
when looking at one’s own batches, where there are buttons to control
running the batch below the pagination buttons.