Add link to edit groups
Pintoch has kindly added support for QuickCategories to the new
EditGroups for Commons already [1], so let’s make our batches link to
the edit groups tool. This is implemented configurably: the config
specifies for each supported domain the URL pattern for the edit group
and (optionally) a date since which batches should link to edit groups
(according to the batch’s creation time).

[1]: https://commons.wikimedia.org/wiki/Commons_talk:Edit_groups/Adding_a_tool#Adding_QuickCategories_support