Fix empty titles in runner
These showed up in batch #5147 [1] and were forcing the background
runner into a crash loop, at 21487 restarts to date 😬 since the API
doesn’t return any pages for an empty title, treat them like invalid
titles, which I think makes sense.

We should probably disallow these at batch parse time too, but given
that we already have an existing batch with this data, we need to handle
it in the background runner anyway (unless I want to go edit the batch
in the database).

In the tests, the “no 'query' in response at all” test case needs the
ability to return different responses based on the request (because the
tokens and siteinfo requests still need a 'query' in the response), so
add the ability for a Callable response (factory) to test_utils.

[1]: https://quickcategories.toolforge.org/batch/5147/