Disable querytime mechanism by default
This writes a lot of data to the database, for little usefulness in
return: I can trace the optimization of 7028c292e7 (“lock less tables”)
to this, but overall I don’t look at the data very much. So let’s stop
writing querytime by default, but make it configurable in the config, by
having DatabaseStore read an 'enable_querytime' param from its
connection params (defaults to false).

Test this in test_querytime; arguably it fits better in test_database,
but test_querytime is where we already have a clean_querytime fixture.