Remove some wmflabs references
Update the ToolsDB host name to the current recommendation; remove the
referrer check in submitted_request_valid(), which (I think) hasn’t been
needed since the toolforge.org migration; update the motivation for
denying iframes (which we continue doing); and update the Toolforge link
(toolforge.org just redirects to a Wikitech page, but let’s still link
to it in case it changes later).

tools-static (for cdnjs) and the Toolforge Docker registry are still on
tools.wmflabs.org as far as I can tell.