Support ProofreadPage index pages
Used in batch #6794 [1] and previously sent the runner into a crash
loop; seems to work fine [1] if we send the right contentmodel with the
API request. Also improve the error message for future situations.

Test this on test2wiki, which (unlike testwiki) has ProofreadPage
installed, using some new test pages. (Note that the new test pages have
spaces instead of slashes to separate the CI job number from the rest of
the title, because with a slash in the title it’s apparently not a
ProofreadPage index page anymore.)

[1]: https://quickcategories.toolforge.org/batch/6794/
[2]: https://ta.wikisource.org/w/index.php?oldid=1525136&diff=1526200