Clean up whitespace more thoroughly
[[User:Strainu]] [1] informed me [2] that MediaWiki treats
[[Category:Bla]] and [[Category: Bla]] as equivalent; closer inspection
of the source code reveals that MediaWikiTitleCodec::splitTitleString()
treats a whole set of characters as equivalent to space/underscore,
trims them from the start and end of the string, and also collapses
consecutive spaces to single ones. Handle all that when looking for
existing categories (to avoid re-adding them, or to correctly remove
them).

Note that we don’t clean up such whitespace in submitted batches; at the
moment, it should get added as the user entered it (though I haven’t
tested this, to be honest). Maybe I’ll add it to cleanup() later (where
we currently just replace underscores with spaces, but don’t do any
other replacements nor trimming), but Strainu says that some users
prefer the form with the leading space, so let’s let them add it this
way for the time being.

(It’s possible that code like _clean_up_whitespace() exists somewhere in
mwparserfromhell, but if it does I can’t find it. Hopefully it won’t
change too much over time.)

[1]: https://meta.wikimedia.org/wiki/User:Strainu
[2]: https://meta.wikimedia.org/wiki/User_talk:Lucas_Werkmeister/QuickCategories#Trimming_category_names