Extract configuration setup into new init module
This consolidates some duplicate code between app.py and
background_runner.py; it’s fairly simple for now, but I’m planning to
extend it to let the app be configured via environment variables (aka
toolforge envvars [1]), so it’s nice to have one place for that upcoming
code. (config.yaml support will almost certainly be kept for local
development, by the way.)

Note that read_private() can be replaced with toolforge’s
load_private_yaml() [2] once a new version of that library is released.

(Also, I considered calling the new module setup.py but I don’t think
Setuptools is sufficiently obsolete for that not to be confusing yet ^^)

[1]: https://wikitech.wikimedia.org/wiki/Help:Toolforge/Envvars
[2]: https://gitlab.wikimedia.org/toolforge-repos/python-toolforge/-/merge_requests/21