Load config from environment
Configuration is now additionally loaded from environment variables
starting with TOOL_*; for instance, the secret key can be set via
TOOL_SECRET_KEY. Nested keys can be set with two underscores, e.g.
TOOL_OAUTH__consumer_key for OAUTH.consumer_key.

The intended goal for this is to make config.yaml obsolete in production
(though it will remain supported for local development) and only
configure the tool via toolforge envvars, as a prerequisite for porting
the tool to using the build service. However, for that to be fully
possible I still need to add a follow-up to read the database
credentials from TOOL_TOOLSDB_USER and TOOL_TOOLSDB_PASSWORD (right now
the environment variables will end up in the config but not be used).