Add support for reading toolsdb from envvars
Previously, we passed the DATABASE config directly into
pymysql.connect() and used the read_default_file= parameter to read the
username and password from ~/replica.my.cnf. This isn’t going to work in
a build pack image (I assume), where we instead want to read the
username and password from environment variables; so let’s add a bit of
custom code to populate user= and password= from there, and also set a
default host= while we’re at it.