Add health check to background runner
This should automatically restart the background runner if it gets
stuck, as it is sometimes wont to do. Touch a file each time we check
for a new command to run, and remove that file in the health check. (We
don’t need an extra `test -e` as part of the health check, as suggested
by the Toolforge documentation – unlink’s exit status will already tell
us if the file existed or not.)

Shorten the sleep time when no command was found from ten to five
seconds. The health check runs every ten seconds and restarts the job if
it fails three times in a row, so sleeping for ten seconds makes me a
bit nervous.

Bug: T374152