Clean up tests for invalid datetime_to_utc_timestamp()
These were raising a deprecation warning for datetime.utcnow(). In
addition, it turns out that replacing utcnow() with now(UTC) did not
fail the test even though it should, because the AssertionError would
still be raised (most of the time) due to the nonzero microseconds in
the datetime. So split the test up: one for the invalid timezone
(setting the microseconds to 0 to ensure that assertion isn’t triggered)
and one for the invalid microseconds. And drop the utcnow() test case –
it doesn’t add anything over the now() test case (both give you a naive
datetime, it doesn’t matter that one of them has the numbers
corresponding to local time and the other to UTC).