Rename *Store to *BatchStore
I’m about to introduce a second kind of store for preferences, and I
think it makes sense to do that in a separate class but in the same
file, so rename the existing stores to make it clear what they are (and
match the abstract base class).

Also update store in test_store.py accordingly.