Add preference for watchlist behavior
Previously, the tool used the default of watchlist=preferences, i.e.,
watch edited pages iff the user enabled that preference. I’ve heard
feedback (about another tool) that this is not always desirable, so
introduce a preference to let the user choose two alternatives in
addition to the default behavior: always watch pages edited with this
tool, or never update the watchlist. (The API supports a fourth option,
forcibly unwatching edited pages, but I see no reason to present that to
users. But it’s useful for the tests, as the CI bot password doesn’t
have the grant to actively edit the watchlist.) The default behavior is
still to follow the preference, since I think that makes the most sense
and is least surprising, even though others argued the default should be
watchlist=nochange [1].

As this preference must be available in the background runner, we can’t
store it in the flask session. Instead, introduce a new database table
for preferences, and a new store for accessing them there. Since the
background runner will have to read this database table once per edit, I
want to optimize its size so MariaDB can cache it in RAM better, so make
both the key and value ints with hard-coded meanings in database.py and
store.py respectively, and access the preference via dedicated methods
instead of general-purpose get/set methods dealing in strings. Other
preferences currently in the session, which I’ll probably move to the
database later, are: whether to show notifications (boolean, trivially
representable as int); the default domain (single optional string,
already denormalized in the domain table, so we can store its ID as an
int); and the suggested domains (list of strings – I guess this will
require a separate table as the preference_value can’t directly hold a
list of domain ID ints, but I’ll deal with that later).

Note that, to make the test work, the bot password will need to have the
“view your watchlist” grant enabled.

[1]: https://meta.wikimedia.org/wiki/User_talk:Lucas_Werkmeister/QuickCategories#Request_for_comments:_Watchlist_behavior