Cache database connection in application context
Previously, loading a batch page would open and close five connections
in rapid succession, and this would’ve gotten worse if I moved e.g. the
“notifications” preference from the flask.session to the database. We
can avoid this by stashing the connection in the Flask application
context and closing it as the context is torn down at the end of each
request. (Note that the database store must also be available in the
background runner, where there is no Flask context, so the app is
optional and we don’t bypass Flask if it’s None.)

In general, connection pooling is not allowed on ToolsDB [1], but I’ve
been told [2] that reusing connections during a single web request
should be okay.

[1]: https://wikitech.wikimedia.org/wiki/Help:Toolforge/Database#Connection_handling_policy
[2]: https://wm-bot.wmcloud.org/browser/index.php?start=12%2F15%2F2024&end=12%2F16%2F2024&display=%23wikimedia-cloud